+++ tests/unit/jquery/dropdown.js
@@ -168,7 +168,7 @@ TestUtil.createJQueryTest(
 );
 
 TestUtil.createJQueryTest(
-    'jquery/dropdown: test load more',
+    'jquery/dropdown: test load more with single input',
     ['inputs/single', 'dropdown', 'templates'],
     function(test, $input, $) {
         $input.selectivity({ query: query });
@@ -200,6 +200,80 @@ TestUtil.createJQueryTest(
 
         test.equal($('.selectivity-result-item').first().text(), 'Amsterdam');
         test.equal($('.selectivity-result-item').last().text(), 'Helsinki');
+    }
+);
+
+TestUtil.createJQueryTest(
+    'jquery/multiple: test load more with multiple input',
+    ['inputs/multiple', 'dropdown', 'templates'],
+    function(test, $input, $) {
+        var offset = null;
+        var term = null;
+
+        function query(queryOptions) {
+            offset = queryOptions.offset;
+            term = queryOptions.term;
+
+            queryOptions.callback({
+                results: ['Amsterdam', 'Antwerp', 'Athens'],
+                more: true
+            });
+        }
+
+        $input.selectivity({ multiple: true, query: query });
+
+        TestUtil.simulateEvent($input[0], 'click');
+
+        test.equal($('.selectivity-dropdown').length, 1,
+                   'Dropdown should\'ve opened');
+        test.equal($('.selectivity-result-item').length, 3,
+                   'Three result items should be displayed');
+        test.equal($('.selectivity-load-more').length, 1,
+                   'The load more option should be present');
+        test.equal(offset, 0,
+                   'The specified offset should be 0');
+        test.equal(term, '',
+                   'The specified search term should be the empty string');
+
+        TestUtil.simulateEvent('.selectivity-load-more', 'click');
+
+        test.equal($('.selectivity-result-item').length, 6,
+                   'Six result items should be displayed');
+        test.equal($('.selectivity-load-more').length, 1,
+                   'The load more option should still be present');
+        test.equal(offset, 3,
+                   'The specified offset should be 3');
+        test.equal(term, '',
+                   'The specified search term should still be the empty string');
+
+        TestUtil.simulateEvent('.selectivity-result-item[data-item-id="Amsterdam"]', 'click');
+
+        test.equal($('.selectivity-dropdown').length, 0,
+                   'Dropdown should\'ve closed');
+
+        TestUtil.simulateEvent($input[0], 'click');
+
+        test.equal($('.selectivity-dropdown').length, 1,
+                   'Dropdown should\'ve opened again');
+        test.equal($('.selectivity-result-item').length, 2,
+                   'Two result items should be displayed (Amsterdam is now filtered)');
+        test.equal($('.selectivity-load-more').length, 1,
+                   'The load more option should be present again');
+        test.equal(offset, 0,
+                   'The specified offset should be 0');
+        test.equal(term, '',
+                   'The specified search term should be the empty string');
+
+        TestUtil.simulateEvent('.selectivity-load-more', 'click');
+
+        test.equal($('.selectivity-result-item').length, 4,
+                   'Four result items should be displayed');
+        test.equal($('.selectivity-load-more').length, 1,
+                   'The load more option should still be present again');
+        test.equal(offset, 3,
+                   'The specified offset should be 3 again');
+        test.equal(term, '',
+                   'The specified search term should still be the empty string');
     }
 );